create application base class for de-duplication
all of the changes in this commit should be mechanical and identical
