+++ lib/chef/application/solo.rb
@@ -16,103 +16,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require_relative "base"
 require_relative "../../chef"
-require_relative "../application"
 require_relative "client"
-require_relative "../client"
-require_relative "../config"
-require_relative "../daemon"
-require_relative "../log"
-require_relative "../config_fetcher"
 require "fileutils" unless defined?(FileUtils)
-require_relative "../mixin/shell_out"
 require "pathname" unless defined?(Pathname)
-require "chef-config/mixin/dot_d"
-require "mixlib/archive" unless defined?(Mixlib::Archive)
-require_relative "../dist"
-require "license_acceptance/cli_flags/mixlib_cli"
 
-class Chef::Application::Solo < Chef::Application
-  include Chef::Mixin::ShellOut
-  include ChefConfig::Mixin::DotD
-  include LicenseAcceptance::CLIFlags::MixlibCLI
+class Chef::Application::Solo < Chef::Application::Base
 
   option :config_file,
     short: "-c CONFIG",
     long: "--config CONFIG",
     default: Chef::Config.platform_specific_path("#{Chef::Dist::CONF_DIR}/solo.rb"),
     description: "The configuration file to use."
-
-  option :config_option,
-    long: "--config-option OPTION=VALUE",
-    description: "Override a single configuration option.",
-    proc: lambda { |option, existing|
-      (existing ||= []) << option
-      existing
-    }
-
-  option :formatter,
-    short: "-F FORMATTER",
-    long: "--format FORMATTER",
-    description: "The output format to use.",
-    proc: lambda { |format| Chef::Config.add_formatter(format) }
-
-  option :force_logger,
-    long: "--force-logger",
-    description: "Use logger output instead of formatter output.",
-    boolean: true,
-    default: false
-
-  option :force_formatter,
-    long: "--force-formatter",
-    description: "Use formatter output instead of logger output.",
-    boolean: true,
-    default: false
-
-  option :profile_ruby,
-    long: "--[no-]profile-ruby",
-    description: "Dump complete Ruby call graph stack of entire #{Chef::Dist::PRODUCT} run (expert only).",
-    boolean: true,
-    default: false
-
-  option :color,
-    long: "--[no-]color",
-    boolean: true,
-    default: true,
-    description: "Use colored output, defaults to enabled."
-
-  option :log_level,
-    short: "-l LEVEL",
-    long: "--log_level LEVEL",
-    description: "Set the log level (auto, trace, debug, info, warn, error, fatal).",
-    proc: lambda { |l| l.to_sym }
-
-  option :log_location,
-    short: "-L LOGLOCATION",
-    long: "--logfile LOGLOCATION",
-    description: "Set the log file location, defaults to STDOUT - recommended for daemonizing.",
-    proc: nil
-
-  option :help,
-    short: "-h",
-    long: "--help",
-    description: "Show this help message.",
-    on: :tail,
-    boolean: true,
-    show_options: true,
-    exit: 0
-
-  option :user,
-    short: "-u USER",
-    long: "--user USER",
-    description: "User to set privilege to.",
-    proc: nil
-
-  option :group,
-    short: "-g GROUP",
-    long: "--group GROUP",
-    description: "Group to set privilege to.",
-    proc: nil
 
   unless Chef::Platform.windows?
     option :daemonize,
@@ -122,88 +38,14 @@ class Chef::Application::Solo < Chef::Application
       proc: lambda { |p| true }
   end
 
-  option :lockfile,
-    long: "--lockfile LOCKFILE",
-    description: "Set the lockfile location. Prevents multiple #{Chef::Dist::SOLO} processes from converging at the same time.",
-    proc: nil
-
-  option :interval,
-    short: "-i SECONDS",
-    long: "--interval SECONDS",
-    description: "Run #{Chef::Dist::CLIENT} periodically, in seconds.",
-    proc: lambda { |s| s.to_i }
-
-  option :json_attribs,
-    short: "-j JSON_ATTRIBS",
-    long: "--json-attributes JSON_ATTRIBS",
-    description: "Load attributes from a JSON file or URL.",
-    proc: nil
-
-  option :node_name,
-    short: "-N NODE_NAME",
-    long: "--node-name NODE_NAME",
-    description: "The node name for this client.",
-    proc: nil
-
-  option :splay,
-    short: "-s SECONDS",
-    long: "--splay SECONDS",
-    description: "The splay time for running at intervals, in seconds.",
-    proc: lambda { |s| s.to_i }
-
   option :recipe_url,
     short: "-r RECIPE_URL",
     long: "--recipe-url RECIPE_URL",
     description: "Pull down a remote gzipped tarball of recipes and untar it to the cookbook cache."
 
-  option :version,
-    short: "-v",
-    long: "--version",
-    description: "Show #{Chef::Dist::PRODUCT} version.",
-    boolean: true,
-    proc: lambda { |v| puts "#{Chef::Dist::PRODUCT}: #{::Chef::VERSION}" },
-    exit: 0
-
-  option :override_runlist,
-    short: "-o RunlistItem,RunlistItem...",
-    long: "--override-runlist RunlistItem,RunlistItem...",
-    description: "Replace current run list with specified items for a single run.",
-    proc: lambda { |items|
-      items = items.split(",")
-      items.compact.map do |item|
-        Chef::RunList::RunListItem.new(item)
-      end
-    }
-
-  option :client_fork,
-    short: "-f",
-    long: "--[no-]fork",
-    description: "Fork #{Chef::Dist::CLIENT} process."
-
-  option :why_run,
-    short: "-W",
-    long: "--why-run",
-    description: "Enable whyrun mode.",
-    boolean: true
-
   option :ez,
     long: "--ez",
     description: "A memorial for Ezra Zygmuntowicz.",
-    boolean: true
-
-  option :environment,
-    short: "-E ENVIRONMENT",
-    long: "--environment ENVIRONMENT",
-    description: "Set the #{Chef::Dist::PRODUCT} environment on the node."
-
-  option :run_lock_timeout,
-    long: "--run-lock-timeout SECONDS",
-    description: "Set maximum duration to wait for another client run to finish, default is indefinitely.",
-    proc: lambda { |s| s.to_i }
-
-  option :minimal_ohai,
-    long: "--minimal-ohai",
-    description: "Only run the bare minimum Ohai plugins #{Chef::Dist::PRODUCT} needs to function.",
     boolean: true
 
   option :delete_entire_chef_repo,
@@ -215,8 +57,6 @@ class Chef::Application::Solo < Chef::Application
     long: "--legacy-mode",
     description: "Run #{Chef::Dist::SOLO} in legacy mode.",
     boolean: true
-
-  attr_reader :chef_client_json
 
   # Get this party started
   def run(enforce_license: false)
@@ -294,10 +134,6 @@ class Chef::Application::Solo < Chef::Application
     end
   end
 
-  def setup_application
-    Chef::Daemon.change_privilege
-  end
-
   def run_application
     if !Chef::Config[:client_fork] || Chef::Config[:once]
       begin
@@ -366,12 +202,5 @@ class Chef::Application::Solo < Chef::Application
         f.write(r.read)
       end
     end
-  end
-
-  def unforked_interval_error_message
-    "Unforked #{Chef::Dist::CLIENT} interval runs are disabled by default." +
-      "\nConfiguration settings:" +
-      ("\n  interval  = #{Chef::Config[:interval]} seconds" if Chef::Config[:interval]).to_s +
-      "\nEnable #{Chef::Dist::CLIENT} interval runs by setting `:client_fork = true` in your config file or adding `--fork` to your command line options."
   end
 end
